Separate auth0.types from auth0.utils.
