+++ src/lib/operations/create-user/createUser.utils.ts
@@ -1,7 +1,7 @@
 import { CONFIG } from "../../server/config/config.utils";
 import { kmsClient } from "../../utils/kms/kmsClient";
 import { getKeyRingIdFromIdToken, getKeyRingPath, getLocationPath } from "../../utils/kms/google-kms.utils";
-import { IdTokenWithData } from "../../utils/auth/auth0";
+import { IdTokenWithData } from "../../utils/auth/auth0.utils";
 import { CreateUserIdTokenData, LegacyCreateUserIdTokenData } from "./create-user.handler";
 
 export async function createKeyRing(
