+++ src/lib/operations/fetch-import-job/fetch-import-job.handler.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import { ExpressRequestWithToken } from "../../utils/auth/auth0";
+import { ExpressRequestWithToken } from "../../utils/auth/auth0.utils";
 import { logRequestSuccess, logRequestStart } from "../../utils/log/log.utils";
 import { Route } from "../../server/server.constants";
 import { createOrPropagateError } from "../../server/errors/errors.utils";
