+++ src/lib/operations/fetch-import-job/fetch-import-job.ts
@@ -1,4 +1,4 @@
-import { IdTokenWithData } from "../../utils/auth/auth0";
+import { IdTokenWithData } from "../../utils/auth/auth0.utils";
 import { kmsClient } from "../../utils/kms/kmsClient";
 import { getImportJobPath } from "../../utils/kms/google-kms.utils";
 import { FetchImportJobIdTokenData } from "./fetch-import-job.handler";
