+++ src/lib/utils/auth/auth0.utils.ts
@@ -1,4 +1,3 @@
-import express from "express";
 import { CONFIG } from "../../server/config/config.utils";
 import { B64UrlString, ownerToAddress } from "../arweave/arweaveUtils";
 import axios from "axios";
@@ -9,32 +8,7 @@ import { pem2jwk } from "pem-jwk";
 import { getSignKeyVersionPath } from "../kms/google-kms.utils";
 import { kmsClient } from "../kms/kmsClient";
 import { ActivateKeysIdTokenData } from "../../operations/activate-keys/activate-keys.handler";
-import type { IdToken } from "@auth0/auth0-spa-js";
-
-export interface UserMetadata {
-  // Custom from Auth0's Add User Metadata action:
-  owner: B64UrlString; // Public key derived from `sub`.
-  walletAddress: B64UrlString; // Wallet address derived from `owner`.
-  authSystem: "KMS";
-}
-
-
-export interface IdTokenWithData<D = void> extends IdToken, UserMetadata {
-  // Extra data also added to the token in Add User Metadata action when calling functions other than createUser:
-  data: void extends D ? never : D;
-}
-
-export interface ExpressRequestWithToken<D = void> extends express.Request {
-  idToken: IdTokenWithData<D>;
-}
-
-export function isExpressRequestWithToken(
-  req: express.Request | ExpressRequestWithToken,
-): req is ExpressRequestWithToken {
-  return (
-    req.hasOwnProperty("idToken") && !!(req as ExpressRequestWithToken).idToken
-  );
-}
+import { IdTokenWithData, UserMetadata } from "./auth0.types";
 
 export type ValidAuth0Pathnames =
   | "/oauth/token/"
